添加代码生成器，数据库操作的代码再也不手写啦 修复相关的依赖，以及新出的bug

+++ src/main/java/cn/iotlearn/community/dto/PaginationDTO.java
@@ -16,7 +16,7 @@ public class PaginationDTO {
     private Integer totalPage;
     private List<Integer> pages = new ArrayList<>();
 
-    public void setPageination(Integer totalCount, String page, String size) {
+    public void setPageination(long totalCount, String page, String size) {
         int totalPage = (int) Math.ceil((double)totalCount/Integer.parseInt(size));
         this.totalPage = totalPage;
         this.page = Integer.parseInt(page);
